+++ termz/util/debug.py
@@ -41,6 +41,12 @@ from collections.abc import Callable
 from functools import wraps
 from typing import ParamSpec, TypeVar
 
+__all__ = [
+    "print_arguments",
+    "timing",
+    "timing_ns",
+]
+
 _logger = logging.getLogger(__name__)
 
 P = ParamSpec("P")